+++ backend/src/user/models/photo.model.ts
@@ -2,7 +2,6 @@ import { Column, DataType, ForeignKey, Model, Table } from "sequelize-typescript
 import { ApiProperty } from "@nestjs/swagger";
 import { UserModel } from "./user.model";
 import { PostModel } from "src/post/models/post.model";
-import { Post } from "@nestjs/common";
 
 interface PhotoCreationAttr {
   userId: number;